contentflow-api/requirements.txt
# Core dependencies
pydantic>=2.12.5
pydantic-settings==2.12.0
pyyaml>=6.0.3
python-dotenv>=1.2.1
httpx==0.28.1
aiohttp>=3.13.2

# FastAPI framework
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.20

# Azure libraries
azure-core==1.37.0
azure-identity==1.25.1
azure-storage-blob==12.27.1
azure-storage-queue==12.14.1
azure-cosmos==4.14.3
azure.appconfiguration.provider==2.3.1

# Authentication & Security
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4

-e ../contentflow-lib